NEW: split into two pages
- Landing page show stats
- Then you can see nodes list and filter them
