--- app/nodes/template.hbs
@@ -1,75 +1,0 @@
-{{#if model.getNodes.isRunning}}
-  <div class="layout flex layout-align-space-around">
-    {{paper-progress-circular warn=true class="md-hue-3" diameter=96}}
-  </div>
-{{else}}
-  <div class="layout flex layout-align-space-araund md-padding">
-    <div class="flex-55">
-      <img src="/assets/images/logo.png" class="logo" alt="Bitcon Cash"/>
-    </div>
-    <div class="flex-20">
-      <h6 class="powered--text">powered by</h6>
-      <a href="https://www.bitcoinunlimited.info/" target="_blank">
-        <img class="powered--logo" src="https://www.bitcoinunlimited.info/img/bitcoin-unlimited.png" alt="Bitcoin Unlimited"/>
-      </a>
-    </div>
-  </div>
-  <div class="layout-column flex layout-align-space-araund md-padding">
-    <h1>Bitcoin Cash Nodes</h1>
-    <p>
-      Snapshot of reachable Bitcoin Cash nodes as of <strong>{{snapshotDate}}</strong>: total <strong>{{nodesCount}}</strong> nodes.
-    </p>
-    <div class="layout-row md-padding layout-xs-column">
-      <div class="flex">
-        {{pie-chart data=nodesByUserAgentPie options=userAgentPieOptions}}
-      </div>
-      <div class="flex">
-        {{#paper-tabs
-          center=center
-          stretch=stretch
-          borderBottom=borderBottom
-          selected=selectedBasicTab
-          onChange=(action (mut selectedBasicTab)) as |tabs|}}
-          {{#tabs.tab}}
-            User Agent
-          {{/tabs.tab}}
-          {{#tabs.tab}}
-            Countries
-          {{/tabs.tab}}
-          {{#tabs.tab}}
-            Network
-          {{/tabs.tab}}
-        {{/paper-tabs}}
-        {{top-ten-tables
-            tableIdx=selectedBasicTab
-            countriesData=nodesCountByCountry
-            networkData=nodesCountByNetwork
-            userAgentData=nodesCountByUserAgent}}
-      </div>
-    </div>
-    <div class="layout-row md-padding">
-      <div class="flex layout-align-space-around">
-        {{geo-chart data=geoData}}
-      </div>
-    </div>
-    <div class="layout-row md-padding">
-      {{paper-input class="flex-50" label="Filter nodes, e.g., BUCash:1 or, 73.172" value=filter onChange=(action 'filterNodes') icon="search"}}
-    </div>
-    <p>
-      Showing {{fileterNodesCount}} out of {{nodesCount}}.
-    </p>
-    <div class="md-padding">
-      {{occluded-table model=nodesData columns=tableColumns}}
-    </div>
-    <div class="md-padding text-center footer">
-      <p>
-        Start your own Bitcoin Cash full node on your Linux, Mac, BSD or Windows system by running this command:
-      </p>
-      <p class="code">
-        <code>
-          curl <a href=https://cashnodes.io/install-bucash-full-node.sh> https://cashnodes.io/install-bucash-full-node.sh</a> | sh
-        </code>
-      </p>
-    </div>
-  </div>
-{{/if}}
