+++ app/stats/controller.js
@@ -1,0 +1,6 @@
+import Controller from '@ember/controller';
+
+import NodesDataMixin from '../mixins/nodes-data';
+
+export default Controller.extend(NodesDataMixin, {
+});
