+++ app/stats/nodes/route.js
@@ -1,0 +1,8 @@
+import Route from '@ember/routing/route';
+
+export default Route.extend({
+  model() {
+    const nodes = this.modelFor('stats');
+    return nodes;
+  },
+});
