+++ app/stats/nodes/template.hbs
@@ -1,0 +1,9 @@
+<div class="layout-row md-padding">
+  {{paper-input class="flex-50" label="Filter nodes, e.g., BUCash:1 or, 73.172" value=filter onChange=(action 'filterNodes') icon="search"}}
+</div>
+<p>
+  Showing {{fileterNodesCount}} out of {{nodesCount}}.
+</p>
+<div class="md-padding">
+  {{occluded-table model=nodesData columns=tableColumns}}
+</div>
