+++ app/stats/route.js
